# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NBE10-12-2-Team1 API
  description: |
    오늘 뭐먹지 SNS 서비스 REST API 문서입니다.

    ## 인증
    인증이 필요한 API는 요청 헤더에 Bearer 토큰을 포함해야 합니다.
    ```
    Authorization: Bearer {accessToken}
    ```
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/feeds:
    post:
      tags: [Feeds]
      summary: 피드 글 작성
      description: 새로운 피드 글을 작성합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [feed]
              properties:
                feed:
                  type: object
                  required: [content, restaurantId]
                  properties:
                    content:
                      type: string
                      maxLength: 1000
                      example: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                    restaurantId:
                      type: integer
                      example: 1
                image:
                  type: string
                  format: binary
                  description: 피드 이미지 파일 (jpg/jpeg/png, 선택)
      responses:
        "201":
          description: 글 작성 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feedId: 45
                  content: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                  nickname: "푸디보이"
                  profileImage: "https://k.kakaocdn.net/.../img.jpg"
                  likeCount: 0
                  createdAt: "2026-06-18T17:50:00"
                  updatedAt: "2026-06-18T17:50:00"
                  restaurantId: 1
                  restaurantName: "을지삼겹"
                message: "피드가 생성되었습니다."
        "400":
          description: content 없음 또는 1000자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "내용을 입력해주세요. (최대 1000자)"
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "404":
          description: 존재하지 않는 식당 (RESTAURANT_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "RESTAURANT_NOT_FOUND"
                  message: "존재하지 않는 식당입니다."
        "500":
          description: 저장 실패
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    get:
      tags: [Feeds]
      summary: 피드 글 목록 조회
      description: |
        피드 글 목록을 조회합니다.

        **용례:**
        - `GET /api/v1/feeds?restaurantId=1` → 특정 식당 태그된 글들 (식당 페이지)
        - `GET /api/v1/feeds?userId=2` → 특정 유저의 글 (프로필 페이지)
        - `GET /api/v1/feeds` → 전체 공개 글 (탐색)
      parameters:
        - name: restaurantId
          in: query
          required: false
          description: 특정 식당 태그 글만 필터링
          schema:
            type: integer
            example: 1
        - name: userId
          in: query
          required: false
          description: 특정 유저의 글만 필터링
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 글 목록 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feeds:
                    - feedId: 45
                      content: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                      userId: 1
                      nickname: "푸디보이"
                      profileImage: "https://k.kakaocdn.net/.../img.jpg"
                      likeCount: 12
                      commentCount: 4
                      isLikedByMe: true
                      restaurantId: 1
                      restaurantName: "을지삼겹"
                      createdAt: "2026-06-18T17:50:00"
                  totalPages: 3
                  totalElements: 25
                  page: 0
                  size: 10
        "400":
          description: restaurantId/userId 형식 오류 또는 page/size가 0 이하 (INVALID_INPUT_VALUE / INVALID_PAGE_SIZE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "입력값이 올바르지 않습니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/feeds/following:
    get:
      tags: [Feeds]
      summary: 팔로잉 피드 목록 조회
      description: 내가 팔로우한 사용자들의 피드 글 목록을 최신순으로 조회합니다. 로그인 필요.
      parameters:
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 팔로잉 피드 목록 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feeds:
                    - feedId: 45
                      content: "오늘 을지로 삼겹살 먹었는데 대박이야!"
                      userId: 2
                      nickname: "서교동김푸디"
                      profileImage: "https://k.kakaocdn.net/.../profile2.jpg"
                      likeCount: 12
                      commentCount: 4
                      isLikedByMe: false
                      restaurantId: 1
                      restaurantName: "을지삼겹"
                      createdAt: "2026-06-18T17:50:00"
                  totalPages: 3
                  totalElements: 25
                  page: 0
                  size: 10
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/feeds/recommend:
    get:
      tags: [Feeds]
      summary: 추천 피드 목록 조회
      description: 팔로우하지 않은 사용자의 피드 중 랜덤으로 추천 피드 목록을 조회합니다. 로그인 필요.
      parameters:
        - name: page
          in: query
          required: false
          description: 페이지 번호 (기본값 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기 (기본값 10)
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 추천 피드 목록 조회 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feeds:
                    - feedId: 88
                      content: "망원동 숨집 발견! 분위기 미쳤다"
                      userId: 3
                      nickname: "망원동고독가"
                      profileImage: "https://k.kakaocdn.net/.../profile3.jpg"
                      likeCount: 5
                      commentCount: 2
                      restaurantId: 5
                      restaurantName: "망원동숨집"
                      createdAt: "2026-06-19T12:30:00"
                  totalPages: 5
                  totalElements: 48
                  page: 0
                  size: 10
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
  /api/v1/feeds/{id}:
    put:
      tags: [Feeds]
      summary: 피드 글 수정
      description: 피드 글을 수정합니다. 본인만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 글 ID
          schema:
            type: integer
            example: 45
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [feed]
              properties:
                feed:
                  type: object
                  required: [content, restaurantId]
                  properties:
                    content:
                      type: string
                      maxLength: 1000
                      example: "재방문했는데 여전히 맛있네요. 강력 추천합니다."
                    restaurantId:
                      type: integer
                      example: 2
                image:
                  type: string
                  format: binary
                  description: 피드 이미지 파일 (jpg/jpeg/png, 선택)
      responses:
        "200":
          description: 글 수정 성공
          content:
            application/json:
              example:
                success: true
                data:
                  feedId: 45
                  content: "재방문했는데 여전히 맛있네요. 강력 추천합니다."
                  nickname: "푸디보이"
                  profileImage: "https://k.kakaocdn.net/.../img.jpg"
                  likeCount: 12
                  createdAt: "2026-06-18T17:50:00"
                  updatedAt: "2026-06-18T18:10:00"
                  restaurantId: 2
                  restaurantName: "망원동숨집"
                message: "피드가 수정되었습니다."
        "400":
          description: content 1000자 초과
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INVALID_INPUT_VALUE"
                  message: "내용은 1000자를 초과할 수 없습니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 타인 글 수정 시도 (FORBIDDEN_FEED_UPDATE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN_FEED_UPDATE"
                  message: "본인의 글만 수정할 수 있습니다."
        "404":
          description: 존재하지 않는 글 (FEED_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FEED_NOT_FOUND"
                  message: "존재하지 않는 게시글입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."

    delete:
      tags: [Feeds]
      summary: 피드 글 삭제
      description: 피드 글을 삭제합니다. 본인만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 글 ID
          schema:
            type: integer
            example: 45
      responses:
        "200":
          description: 글 삭제 성공
          content:
            application/json:
              example:
                success: true
                data: null
                message: "글이 성공적으로 삭제되었습니다."
        "401":
          description: 로그인하지 않은 사용자
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "UNAUTHORIZED"
                  message: "로그인 후 이용해주세요."
        "403":
          description: 타인 글 삭제 시도 (FORBIDDEN_FEED_DELETE)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FORBIDDEN_FEED_DELETE"
                  message: "본인의 글만 삭제할 수 있습니다."
        "404":
          description: 존재하지 않는 글 (FEED_NOT_FOUND)
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "FEED_NOT_FOUND"
                  message: "존재하지 않는 게시글입니다."
        "500":
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                data: null
                error:
                  code: "INTERNAL_SERVER_ERROR"
                  message: "서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
...